# code/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'REST API'
  version: 1.0.0
paths:
...
  /api/v1/profile/user/password:
    put:
      tags:
        - Profile
      summary: 'Change user password'
      description: 'Changes the password for the authenticated user'
      operationId: e612eff67a2cf0133671fcb0503eb233
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  example: current123
                newPassword:
                  type: string
                  example: new123
              type: object
      responses:
        '200':
          description: 'Password changed successfully'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: Ok }
                type: object
        '400':
          description: 'Validation error'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'Password change failed' }
                type: object
    post:
      tags:
        - Profile
      summary: 'Create user password'
      description: 'Create the password for the authenticated user'
      operationId: 68e4bb483437d99c91fbc985909b1fb8
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - password
                - confirmationPassword
              properties:
                password:
                  type: string
                  example: current123
                confirmationPassword:
                  type: string
                  example: new123
              type: object
      responses:
        '200':
          description: 'Password created successfully'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: Ok }
                type: object
        '400':
          description: 'Validation error'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'Password created failed' }
                type: object
  /api/v1/profile/user/status:
    put:
      tags:
        - Profile
      summary: 'Update user status'
      description: 'Update user account status (activate/deactivate)'
      operationId: f7dc9b813faf06d50d3ec060a7d63576
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                status:
                  description: 'User status'
                  type: string
              type: object
      responses:
        '200':
          description: 'User status updated successfully'
...
  /v1/profile/user:
    put:
      tags:
        - Profile
      summary: 'Update user data'
      description: 'Update user data'
      operationId: e7022ce0348fba50f851eeb152f1a002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                avatar:
                  type: string
              type: object
      responses:
        '200':
          description: 'User updated successfully'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                type: object
        '400':
          description: 'Invalid input or user already exists'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string }
                type: object
        '401':
          description: 'User not authenticated'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string }
                type: object
        '404':
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string }
                type: object
...